requirements.txt
asgiref==3.5.0
Babel==2.10.1
Django==4.0.4
django-crispy-forms==1.14.0
django-environ==0.8.1
django-money==2.1.1
Pillow==9.1.0
py-moneyed==1.2
pytz==2022.1
sqlparse==0.4.2
